+++ www/shiki-twoslash.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  themes: ["min-light", "nord"],
+};
